+++ components/modules/SubsectionHeader.tsx
@@ -1,7 +1,7 @@
 import { TypeAnimation } from "react-type-animation";
 
 interface SubsectionHeaderProps {
-  subtitle: string;
+  subtitle?: string;
   typeSequence: (number | string)[];
 }
 
@@ -11,7 +11,15 @@ const SubsectionHeader = ({
 }: SubsectionHeaderProps) => {
   return (
     <div className="flex flex-col">
-      <div className="text-2xl sm:text-3xl md:text-4xl lg:text-5xl font-bold bg-gradient-to-r from-pink-500 via-purple-500 to-blue-500 text-opacity-0 pb-3 text-white bg-clip-text text-center">
+      <div
+        className={`${
+          subtitle
+            ? "text-2xl sm:text-3xl md:text-4xl"
+            : "text-3xl sm:text-4xl md:text-5xl"
+        } lg:text-5xl font-bold bg-gradient-to-r from-pink-500 via-purple-500 to-blue-500 text-opacity-0 ${
+          subtitle && "pb-3"
+        } text-white bg-clip-text text-center`}
+      >
         <TypeAnimation
           sequence={typeSequence}
           speed={70}
@@ -19,7 +27,11 @@ const SubsectionHeader = ({
           cursor={true}
         />
       </div>
-      <h1 className="text-xl md:text-3xl font-thin text-center">{subtitle}</h1>
+      {subtitle && (
+        <h1 className="text-xl md:text-3xl font-thin text-center">
+          {subtitle}
+        </h1>
+      )}
     </div>
   );
 };