+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 0.1.0a
 - improved contact
 - changed social icons to backtotop
+
+0.1.1
+- made backtotop show on mobile
+- attempt to allow mobile users see hero background video
+- added story page, and moved hobbies into the story page